Add smart setuid wrapper for procmail
Allows us to use a strategy similar to postfix; instead of running
procmail or another LDA with root privileges and trusting it to drop
them before it does anything stupid, run procmail as the user that
is receiving the mail. This allows simta to work out of the box
on distros that have removed the setuid bit from procmail, while
respecting the security concerns that led to that decision (which a
simpler wrapper does not do).