Improve alias handling
Now more compatible with other programs that shall not be named.

Aliases are not required to be on a single line or to be separated
by spaces, so we can't just throw a line at acav_parse() and call it
good. The rewritten parser does not support \ before a newline as
a continuation (there are references to this in documentation for
some MTAs, but it's not universal and I don't like it), but it's more
robust than it used to be.

Add as a special case support for "delivering" mail to /dev/null,
which is a common idiom in alias files. Usually this is a generic
feature that delivers to any file, but I don't want to support that.

Support group owners in alias files. Some MTAs use the convention
"owner-groupname"; we canonicalise this to "groupname-errors" to be
consistent with LDAP.