Rework metric file writing
Older versions of simta created unique files each time, and there are
benefits to that. There are also benefits to having a fixed filename for
the latest file written. So, each time we write a file we write it to a
timestamped file, then create a link for the plain file name if writing
was successful.