Clean up config parsing somewhat
There may have been some grand vision behind the ordering of the
parsing stanzas, but it has been lost to the mists of time and changes
that threw things in at random places. Reorder them alphabetically.

Replace *_OFF and *_ON directives with directives that accept ON or OFF.

Rename some directives (particularly some that are TLS related) for
clarity and consistency.