Clean up .h files
Add multiple inclusion guards and make sure they always include the
headers they depend on. This makes things easier on developers, since
it means that no special knowledge is needed to #include a header.

Clean up some of the unnecessary includes that were only due to header
dependencies and not directly used in that .c file.