Rework header parsing
Centralize the header caching and indexing in header_text(), and
centralize the parsing and logic in header_check().

The mailbox list parsing made some faulty assumptions that often
caused it to fail on legitimate headers, and it heavily intermixed
program logic with parsing; rip it out and replace it with simpler
parsing that is correct in more situations.