Make Command processing more resilient
It should be possible to use simc to promote a queue on any host, not
just a jail. Even if it doesn't work, it should be safe to try. This was
not the case.

  simta: Command C55B7BAD8.30D84.12396: Queue umflint.edu
  simta: Syserror 55B79B04.F1FB6.7670: env_jail_status: illegal env code: 0
  simta: Daemon: Shutdown

Only try to parole messages if we're a jail, and never goto error in
response to a command failing.