Canonicalise BATV addresses during LDAP comparison
Bounce Address Tag Validation (BATV) defines a method for including
tagging information in the local-part of the RFC5321.MailFrom address,
allowing senders to tag all outgoing mail and reject bounces that aren't
to a tagged address.

When simta is checking membersonly and moderated conditions, it should
strip off the BATV info and do the comparison against the original
address.