Copy new expansion D files instead of linking
Queue cleanup uses the Dinode to determine whether messages are the
same and assumes that the same message with multiple expansion levels
means an incomplete expansion. Moderation and group forwarding break
this assumption by adding another expansion level.

Breaking the link between the D files makes the actual behaviour match
the assumed behaviour; the message to the moderators or group archive
is treated as a completely unrelated message.