Replace libsrs2
Reimplementing it is easier than fixing it, because of reasons.

This implementation is gratuitously incompatible in an unimportant
way that shouldn't matter, and one important way that does. SRS was
not designed to handle some valid email addresses, so libsrs2 (and
probably all other implementations) will not properly handle

    "SRS0=+h12n=kt=example.com=ex@mple"@example.org
or
    "SRS1=ufbyy=example.org==+h12n=kt=example.com=ex@mple"@mail.example.org

when they encounter them.