Add correct RFC7208.MAILFROM handling
The RFC7208.MAILFROM identity is not the same as the RFC5321.MailFrom
identity when the return path is null. This is somewhat obscure and
most (all?) large mail providers do not publish SPF records for their
HELO identities, so it was skipped in the initial implementation for
simplicity's sake.

Most of this commit is cosmetic changes to retain and expose the
SPF code's internal state so that DMARC and Authentication-Results can
easily reference the identity that it checked.