Improve RFC 5782 compliance
Reference the RFC in comments instead of text from the (now defunct)
Open Relay Database website.

Use "dnsl" and "DNS list" instead of RBL. RBL, despite common usage,
refers to a specific commercial list. (It's also weird to call an
accept list an RBL, as we have been doing.) As mentioned in the RFC a
better generic form is DNSxL (e.g. DNSBL -> "DNS block list").

In addition to the A record used by ORDB and other early lists, it's now
common to have a corresponding TXT record with a "reason" for the
listing. Add support for retrieving and using this record in SMTP
responses and logging.

Replace hard-coded independent config/extern variables with named
chains, which greatly simplifies config processing and makes it easier
to add additional lookup chains in the future. The old config directives
are still accepted for now, but should be considered deprecated.