Always lock and reread envelopes during delivery
It was possible for n+1 queue runners to be processing the same
message at the same time if n had read the envelope and then requeued
it (e.g. for punting.)

Always locking the envelope is simpler than figuring out in which
circumstances it's safe to leave it unlocked.

Always rereading the recipients avoids the potential for multiple
deliver when one process rewrites the envelope and releases the lock,
then another process with a cached idea of which recipients remain
processes the message.