Revert "invalid character in DATA" check
We have at least two valid sources of mail that are tripping this.
simta will regularize the line endings and do dot stuffing when
delivering the mail, so as long as we have a strict check that the
inbound EOM is `.\r\n` this still avoids the SMTP smuggling attack.